refactor: separate deduplication logic into hook
