--- docs/src/plugins/starlight/mixedbread/virtual.d.ts
@@ -1,10 +1,0 @@
-declare module 'virtual:starlight/mixedbread-config' {
-  const config: {
-    apiKey: string
-    vectorStoreId: string
-    maxResults?: number
-    baseUrl?: string
-    disableUserPersonalization?: boolean
-  }
-  export default config
-}
